FIX The ASR doesn't enforce the grammar
I have observed that for some of the synthesised utterances (possibly
the worst generated ones), the decoded sentence is not six words long,
as it should be. This is not expected since, we had a rigid grammar that
should have been enforced by the model. However, in order to properly
calculate the per-word error, I had to ignore some of the utterances.